requirements.txt
requests
snscrape
vaderSentiment
solana==0.30.0
solders
python-dotenv
base58
websocket-client
certifi==2023.6.15
